Complexity now given as unsigned longs
It was found that the range of integer values was not large enough to hold the complexity of an some sorting algorithms. Unsigned longs were used instead. Next stage is to implement multi-threading on some algorithms to speed up the program. For future reference, for n = 50,000, bubblesort took an average time of 27.21s on my laptop