Continued work on a locking mechanism
Multithreading bubblesort will now use 4 threads (as specified in a macro)

REFACTORING FUNCTIONS
The bubblesort/multithreading functions seem to be getting too tightly coupled with each other. Possible areas to improve this is to restore the function bubblesort to purely bubblesorting and change bubblesortThreadHelper so that it only creates threads. A separate function will take the return value from these threads and determine the total complexity. A new function will be made to control the locking mechanism by monitoring the indices accessed by each iteration of the bubblesort function